refactor(data): move the data layer onto compile-time checked queries
Replaces runtime `sqlx::query`/`query_as::<_, T>` with the `query!`/`query_as!`
macros throughout `src/data/`, dropping `FromRow` derives and the column
nullability the old form could not verify. Outer-join columns now carry explicit
`AS "col?"` assertions, since sqlx over-reports non-null there and an
`Option<T>` field would swallow it until a real NULL panicked.

- `CourseInstructorDetail::course_id` is non-optional now that both queries select it
- extract `StreamSink` so outbound WebSocket sends and their metrics live in one place
- fold the formatter's colored-write helpers into a `Writer` extension trait

+++ src/web/admin/scraper.rs
@@ -340,7 +340,7 @@ pub async fn scraper_subject_detail(
     let results: Vec<SubjectResultEntry> = rows
         .into_iter()
         .map(|row| SubjectResultEntry {
-            id: row.id as i64,
+            id: row.id,
             completed_at: row.completed_at,
             duration_ms: DurationMs::new(row.duration_ms.max(0) as u32),
             success: row.success,
